merge

+++ examples/scripts/docker_run.sh
@@ -3,6 +3,12 @@ set -x
 PROJECT_PATH=$(cd $(dirname $0)/../../; pwd)
 IMAGE_NAME="nvcr.io/nvidia/pytorch:24.02-py3"
 
-docker run --runtime=nvidia -it --rm --shm-size="10g" --cap-add=SYS_ADMIN \
-	-v $PROJECT_PATH:/openrlhf -v  $HOME/.cache:/root/.cache -v  $HOME/.bash_history2:/root/.bash_history \
-	$IMAGE_NAME bash
+if [[ "${build}" == *"b"* ]]; then
+	# for retry
+	# docker image rm $IMAGE_NAME
+	docker build -t $IMAGE_NAME $PROJECT_PATH/dockerfile 
+else 
+	docker run --runtime=nvidia -it --rm --shm-size="10g" --cap-add=SYS_ADMIN \
+		-v $PROJECT_PATH:/openrlhf -v  $HOME/.cache:/root/.cache -v  $HOME/.bash_history2:/root/.bash_history \
+		$IMAGE_NAME bash
+fi
